==> 3-Text_classification.ipynb <==
<font color="grey">
    
## Zero-Shot Classification

- Zero Shot Classification is the task of predicting a class that wasn't seen by the model during training. This method, which leverages a pre-trained language model, can be thought of as an instance of transfer learning which generally refers to using a model trained for one task in a different application than what it was originally trained for. This is particularly useful for situations where the amount of labeled data is small.
- In zero shot classification, we provide the model with a prompt and a sequence of text that describes what we want our model to do, in natural language. Zero-shot classification excludes any examples of the desired task being completed. This differs from single or few-shot classification, as these tasks include a single or a few examples of the selected task.

Learning links:
- https://huggingface.co/tasks/zero-shot-classification
- https://huggingface.co/models
- https://huggingface.co/models?pipeline_tag=zero-shot-classification&sort=trending

</font>

import pandas as pd
books=pd.read_csv(r'F:\learning\LLM\data\books_cleaned.csv')

books.columns

books['categories'].value_counts().reset_index()

books['categories'].value_counts().reset_index().query('count>50')

category_mapping = {'Fiction': "Fiction",
                    'Juvenile Fiction': "Fiction",
                    'Biography & Autobiography': "Nonfiction",
                    'History': "Nonfiction",
                    'Literary Criticism': "Nonfiction",
                    'Philosophy': "Nonfiction",
                    'Religion': "Nonfiction",
                    'Comics & Graphic Novels': "Fiction",
                    'Drama': "Fiction",
                    'Juvenile Nonfiction': "Nonfiction",
                    'Science': "Nonfiction",
                    'Poetry': "Fiction"}
books['simple_categories']=books["categories"].map(category_mapping)

print("The proportion of books with missing categories:")
round((books['simple_categories'].isna().sum())/len(books),2).item()

<font color="grey">
    
More models about zero-shot-classification:
- https://huggingface.co/models?pipeline_tag=zero-shot-classification&sort=trending

</font>

from transformers import pipeline
fiction_categories = ["Fiction", "Nonfiction"]
pipe = pipeline("zero-shot-classification", model="facebook/bart-large-mnli")
#prediction=pipe(text_description, category space)

book1=books.loc[books['simple_categories']=='Fiction','description'][0]
book1

import numpy as np
prediction=pipe(book1, ["Fiction", "Nonfiction"])
print(f"Prediction keys: {prediction.keys()}")
print(f"Prediction labels: {prediction['labels']}")
print(f"Prediction scores: {prediction['scores']}")
print(f"The predicted category: {prediction['labels'][np.argmax(prediction['scores'])]}")

def generate_predictions(sequence, categories):
    prediction=pipe(sequence, categories)
    max_index=np.argmax(prediction['scores'])
    max_label=prediction['labels'][max_index]
    return max_label # Predicted category

<font color="grey">
    
In our database, there are 2 types of books: those with existing category information and those without it. Approximately 30% of the books lack category information.

We first evaluate the model using books that already have category labels to assess its accuracy. If the performance is satisfactory, we can then use the model to infer category information for books that do have have it.
</font>

from tqdm import tqdm
book_fiction=books.loc[books['simple_categories']=='Fiction','description'].reset_index(drop=True)
book_nonfiction=books.loc[books['simple_categories']=='Nonfiction','description'].reset_index(drop=True)
k=0
predicred_cats=[]
for i in tqdm(range(300)):
    sequence=book_fiction[i]
    pre=generate_predictions(sequence, ["Fiction", "Nonfiction"])
    predicred_cats += [pre]
    if pre=="Fiction":k=k+1     
for i in tqdm(range(300)):
    sequence=book_nonfiction[i]
    pre=generate_predictions(sequence, ["Fiction", "Nonfiction"])
    predicred_cats += [pre]
    if pre=="Nonfiction":k=k+1
actual_cats=["Fiction"]*300+["Nonfiction"]*300
print(f'Text classification accuracy: {k/len(predicred_cats)}')

- 75% is quite good accuracy.

predicted_cats=[]
books_missing_cats=books.loc[books['simple_categories'].isna(),['isbn13','description']].reset_index(drop=True)
for i in tqdm(range(len(books_missing_cats))):
    sequence=books_missing_cats['description'][i]
    predicted_cats += [generate_predictions(sequence, ["Fiction", "Nonfiction"])]

missing_predictions_df=pd.DataFrame({"isbn13":books_missing_cats['isbn13'].values,"predicred_cats":predicted_cats})
books=pd.merge(books,missing_predictions_df,on='isbn13',how='left')
books['simple_categories']=np.where(books['simple_categories'].isna(),books['predicred_cats'],books['simple_categories'])
books=books.drop(columns=['predicred_cats'])
books.to_csv(r"F:\learning\LLM\step3_books_with_categories.csv", index=False)

<font color="grey">
    
### The whole code
</font>

import pandas as pd
import numpy as np
from transformers import pipeline
books=pd.read_csv(r'F:\learning\LLM\data\books_cleaned.csv')
category_mapping = {'Fiction': "Fiction",
                    'Juvenile Fiction': "Fiction",
                    'Biography & Autobiography': "Nonfiction",
                    'History': "Nonfiction",
                    'Literary Criticism': "Nonfiction",
                    'Philosophy': "Nonfiction",
                    'Religion': "Nonfiction",
                    'Comics & Graphic Novels': "Fiction",
                    'Drama': "Fiction",
                    'Juvenile Nonfiction': "Nonfiction",
                    'Science': "Nonfiction",
                    'Poetry': "Fiction"}
books['simple_categories']=books["categories"].map(category_mapping)

pipe = pipeline("zero-shot-classification", model="facebook/bart-large-mnli")
#prediction=pipe(text_description, category space)

def generate_predictions(sequence, categories):
    prediction=pipe(sequence, categories)
    max_index=np.argmax(prediction['scores'])
    max_label=prediction['labels'][max_index]
    return max_label # Predicted category

# We first evaluate the model using books that already have category labels to assess its accuracy
from tqdm import tqdm
book_fiction=books.loc[books['simple_categories']=='Fiction','description'].reset_index(drop=True)
book_nonfiction=books.loc[books['simple_categories']=='Nonfiction','description'].reset_index(drop=True)
k=0
predicred_cats=[]
for i in tqdm(range(300)):
    sequence=book_fiction[i]
    pre=generate_predictions(sequence, ["Fiction", "Nonfiction"])
    predicred_cats += [pre]
    if pre=="Fiction":k=k+1     
for i in tqdm(range(300)):
    sequence=book_nonfiction[i]
    pre=generate_predictions(sequence, ["Fiction", "Nonfiction"])
    predicred_cats += [pre]
    if pre=="Nonfiction":k=k+1
actual_cats=["Fiction"]*300+["Nonfiction"]*300
print(f'Text classification accuracy: {k/len(predicred_cats)}')
   
# Use the model to infer category information for books that do have have it
predicted_cats=[]
books_missing_cats=books.loc[books['simple_categories'].isna(),['isbn13','description']].reset_index(drop=True)
for i in tqdm(range(len(books_missing_cats))):
    sequence=books_missing_cats['description'][i]
    predicted_cats += [generate_predictions(sequence, ["Fiction", "Nonfiction"])]

missing_predictions_df=pd.DataFrame({"isbn13":books_missing_cats['isbn13'].values,"predicred_cats":predicted_cats})
books_o1=pd.merge(books,missing_predictions_df,on='isbn13',how='left')
books_o1['simple_categories']=np.where(books_o1['simple_categories'].isna(),books_o1['predicred_cats'],books_o1['simple_categories'])
books_o1=books_o1.drop(columns=['predicred_cats'])
books_o1.to_csv(r"F:\learning\LLM\step3_books_with_categories.csv", index=False)
