==> 4-sentiment_analysis.ipynb <==
<font color="grey">
    
# classify emotions in English text data

## predicts Ekman's 6 basic emotions, plus a neutral class:
anger 🤬
disgust 🤢
fear 😨
joy 😀
neutral 😐
sadness 😭
surprise 😲

More about the model: https://huggingface.co/j-hartmann/emotion-english-distilroberta-base

</font>

# Load the data
import pandas as pd
from tqdm import tqdm
books=pd.read_csv(r"F:\learning\LLM\step3_books_with_categories.csv")

# Load the model
from transformers import pipeline
classifier = pipeline("text-classification", model="j-hartmann/emotion-english-distilroberta-base", top_k=None)

# Try the model with a simple example
out=classifier("I love this!")[0]
print(pd.DataFrame(out))

# Try one book in our book database
print(f"The world lenth: {len(books['description'][0].split())}")
out=classifier(books['description'][0])[0]
print("The predicted emotions:")
print(pd.DataFrame(out))

# Try another book with very long text.
o1=books['description'].apply(lambda x: len(x.split()))
book_id=o1.idxmax()
print(f"The world lenth: {len((books['description'][book_id]).split())}")
out=classifier(books['description'][book_id])[0]
print("The predicted emotions:")
print(pd.DataFrame(out))

<font color="grey">
    
- We encounter an error because the book's text description exceeds the model's processing limits.

We will split each book's text description into individual sentences and apply the model to each sentences. The aggregated emotion predictions across all sentences will be used as the book's overall emotion.

</font>

def sentiment_each_book(text_des):
    out=classifier(text_des.split("."))#sentense by sentense
    df=pd.DataFrame(out[0]).set_index('label')
    for i in range(1,len(out)):
        dfi=pd.DataFrame(out[i]).set_index('label')
        df=pd.concat([df,dfi],axis=1)
    return df.max(axis=1)

def sentiment_all_books(books):
    df=sentiment_each_book(books['description'][0])
    df.name=books['isbn13'][0]
    for i in tqdm(range(1,len(books))):
        dfi=sentiment_each_book(books['description'][i])
        dfi.name=books['isbn13'][i]
        df=pd.concat([df,dfi],axis=1)
    df=df.T.reset_index().rename(columns={'index':'isbn13'})
    return df
books_with_sentiment=sentiment_all_books(books)    
books_emotion=pd.merge(books,books_with_sentiment,on='isbn13')
books_emotion.to_csv(r"F:\learning\LLM\step4_books_with_emotions.csv", index = False)

<font color="grey">
    
## The whole code
</font>

import pandas as pd
from tqdm import tqdm
books=pd.read_csv(r"F:\learning\LLM\step3_books_with_categories.csv")

from transformers import pipeline
classifier = pipeline("text-classification", model="j-hartmann/emotion-english-distilroberta-base", top_k=None)
out=classifier("I love this!")[0]
print(pd.DataFrame(out))

def sentiment_each_book(text_des):
    out=classifier(text_des.split("."))#sentense by sentense
    df=pd.DataFrame(out[0]).set_index('label')
    for i in range(1,len(out)):
        dfi=pd.DataFrame(out[i]).set_index('label')
        df=pd.concat([df,dfi],axis=1)
    return df.max(axis=1)

def sentiment_all_books(books):
    df=sentiment_each_book(books['description'][0])
    df.name=books['isbn13'][0]
    for i in tqdm(range(1,len(books))):
        dfi=sentiment_each_book(books['description'][i])
        dfi.name=books['isbn13'][i]
        df=pd.concat([df,dfi],axis=1)
    df=df.T.reset_index().rename(columns={'index':'isbn13'})
    return df
    
books_with_sentiment=sentiment_all_books(books)    
books_emotion=pd.merge(books,books_with_sentiment,on='isbn13')
books_emotion.to_csv(r"F:\learning\LLM\step4_books_with_emotions.csv", index = False)
